// src/app/cars/domain/models/car.model.ts
export interface Car {
  id?: number;
  brand: string;
  model: string;
  color: string;
  year: number;
  price: number;
  transmission: string;
  engine: string;
  mileage: number;
  doors: number;
  plate: string;
  location: string;
  description: string;
  fuel: string;
  speed: number;
  image: string;
  images: string[];
}